Add files via upload
YGeneric Canonical Ubuntu media (some OEM Dell media) or other Linux media does not boot to the live environment or installer as usual and instead, the system will hang or get stuck.
The newest kernels have moved the video mode setting into the kernel. So all the programming of the hardware-specific clock rates and registers on the video card happens in the kernel rather than in the X driver when the X server starts. This makes it possible to have high-resolution nice looking splash (boot) screens and flicker-free transitions from boot splash to login screen. Unfortunately, on some cards, this doesn't work properly and you end up with a black screen. Adding the nomodeset parameter instructs the kernel to not load video drivers and use BIOS modes instead until X is loaded.